# Optimize fenwick tree manipulations