--- a/Gopkg.lock
+++ b/Gopkg.lock
@@ -2,18 +2,15 @@
 
 
 [[projects]]
+  digest = "1:cf63454c1e81409484ded047413228de0f7a3031f0fcd36d4e1db7620c3c7d1b"
   name = "github.com/leesper/go_rng"
   packages = ["."]
+  pruneopts = ""
   revision = "5344a9259b21627d94279721ab1f27eb029194e7"
 
 [[projects]]
-  name = "github.com/yourbasic/fenwick"
-  packages = ["."]
-  revision = "7cb325001daae879940edf7b19da8557a51ca5f7"
-  version = "1.2.0"
-
-[[projects]]
   branch = "master"
+  digest = "1:ad6d9b2cce40c7c44952d49a6a324a2110db43b4279d9e599db74e45de5ae80c"
   name = "gonum.org/v1/gonum"
   packages = [
     "blas",
@@ -28,13 +25,17 @@
     "lapack/gonum",
     "lapack/lapack64",
     "mat",
-    "stat"
+    "stat",
   ]
+  pruneopts = ""
   revision = "f0982070f509ee139841ca385c44dc22a77c8da8"
 
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "8b5c3547492f54e6308956cf6dcb413342643e051eaa84297c64fc334646ecf7"
+  input-imports = [
+    "github.com/leesper/go_rng",
+    "gonum.org/v1/gonum/stat",
+  ]
   solver-name = "gps-cdcl"
   solver-version = 1