--- a/internal/fenwick/README.md
+++ b/internal/fenwick/README.md
@@ -1,0 +1,1 @@
+This is https://github.com/yourbasic/fenwick with int64 replaced with uint32 to avoid constructing temporary []int64 slice just for the lib.