There are 2 separate changes:

1. Create bigger initial fenwick tree so we don't need to rebuild it on every add

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkAddMultipleTimes/10-4        6903          11206         +62.34%
BenchmarkAddMultipleTimes/100-4       100042        87661         -12.38%
BenchmarkAddMultipleTimes/1000-4      1581666       1184668       -25.10%
BenchmarkAddMultipleTimes/10000-4     6562377       5292517       -19.35%

benchmark                             old allocs     new allocs     delta
BenchmarkAddMultipleTimes/10-4        35             7              -80.00%
BenchmarkAddMultipleTimes/100-4       305            7              -97.70%
BenchmarkAddMultipleTimes/1000-4      1400           7              -99.50%
BenchmarkAddMultipleTimes/10000-4     2285           7              -99.69%

benchmark                             old bytes     new bytes     delta
BenchmarkAddMultipleTimes/10-4        13696         28816         +110.40%
BenchmarkAddMultipleTimes/100-4       100096        28816         -71.21%
BenchmarkAddMultipleTimes/1000-4      1878176       28816         -98.47%
BenchmarkAddMultipleTimes/10000-4     4976000       28816         -99.42%
```

When number of adds is small the benchmark becomes slower because tree is bigger and number of rebuilds is not high. But as number of adds grows benchmark improves a lot.

2. Fork fenwick tree implementation to use uint32 instead of int64 to avoid constructing extra array to pass it to the fenwick lib

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkAddMultipleTimes/10-4        11206         8562          -23.59%
BenchmarkAddMultipleTimes/100-4       87661         74857         -14.61%
BenchmarkAddMultipleTimes/1000-4      1184668       1104090       -6.80%
BenchmarkAddMultipleTimes/10000-4     5292517       5050204       -4.58%

benchmark                             old allocs     new allocs     delta
BenchmarkAddMultipleTimes/10-4        7              6              -14.29%
BenchmarkAddMultipleTimes/100-4       7              6              -14.29%
BenchmarkAddMultipleTimes/1000-4      7              6              -14.29%
BenchmarkAddMultipleTimes/10000-4     7              6              -14.29%

benchmark                             old bytes     new bytes     delta
BenchmarkAddMultipleTimes/10-4        28816         16528         -42.64%
BenchmarkAddMultipleTimes/100-4       28816         16528         -42.64%
BenchmarkAddMultipleTimes/1000-4      28816         16528         -42.64%
BenchmarkAddMultipleTimes/10000-4     28816         16528         -42.64%
```

The cumulative result of 2 commits is following:

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkAddMultipleTimes/10-4        6903          8562          +24.03%
BenchmarkAddMultipleTimes/100-4       100042        74857         -25.17%
BenchmarkAddMultipleTimes/1000-4      1581666       1104090       -30.19%
BenchmarkAddMultipleTimes/10000-4     6562377       5050204       -23.04%

benchmark                             old allocs     new allocs     delta
BenchmarkAddMultipleTimes/10-4        35             6              -82.86%
BenchmarkAddMultipleTimes/100-4       305            6              -98.03%
BenchmarkAddMultipleTimes/1000-4      1400           6              -99.57%
BenchmarkAddMultipleTimes/10000-4     2285           6              -99.74%

benchmark                             old bytes     new bytes     delta
BenchmarkAddMultipleTimes/10-4        13696         16528         +20.68%
BenchmarkAddMultipleTimes/100-4       100096        16528         -83.49%
BenchmarkAddMultipleTimes/1000-4      1878176       16528         -99.12%
BenchmarkAddMultipleTimes/10000-4     4976000       16528         -99.67%
```

So it is only slightly slower for small number of inserts and if you chose smaller compression that should not matter at all.

In my projects the speedup is measurable too - from 16 seconds to 6 seconds.