@vmihailenco 2018-11-01
For the record this is the benchmark for compression=20 which is just smaller initial fenwick tree:

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkAddCompression1-4            162           164           +1.23%
BenchmarkAddCompression10-4           204           205           +0.49%
BenchmarkAddCompression100-4          268           267           -0.37%
BenchmarkAddMultipleTimes/10-4        4486          3925          -12.51%
BenchmarkAddMultipleTimes/100-4       59962         35770         -40.35%
BenchmarkAddMultipleTimes/1000-4      431884        303813        -29.65%
BenchmarkAddMultipleTimes/10000-4     2800701       2416271       -13.73%

benchmark                             old allocs     new allocs     delta
BenchmarkAddCompression1-4            0              0              +0.00%
BenchmarkAddCompression10-4           0              0              +0.00%
BenchmarkAddCompression100-4          0              0              +0.00%
BenchmarkAddMultipleTimes/10-4        36             7              -80.56%
BenchmarkAddMultipleTimes/100-4       213            7              -96.71%
BenchmarkAddMultipleTimes/1000-4      438            7              -98.40%
BenchmarkAddMultipleTimes/10000-4     641            7              -98.91%

benchmark                             old bytes     new bytes     delta
BenchmarkAddCompression1-4            0             0             +0.00%
BenchmarkAddCompression10-4           0             0             +0.00%
BenchmarkAddCompression100-4          4             0             -100.00%
BenchmarkAddMultipleTimes/10-4        4112          3744          -8.95%
BenchmarkAddMultipleTimes/100-4       45104         3744          -91.70%
BenchmarkAddMultipleTimes/1000-4      183056        3744          -97.95%
BenchmarkAddMultipleTimes/10000-4     393456        3744          -99.05%
```