@caio 2018-11-01
Whoa this is awesome, thanks a lot for the patches and well detailed PR!

I'll make time to look at it asap (this Saturday, most likely), but from a quick look it looks great already