@vmihailenco 2018-11-02
:+1: 

Overall my understanding is that disabling fenwick tree until t-digest is saturated (number of centroids is close to maximum allowed by compression) going to be beneficial in most workloads. In other cases we still need to rebuild fenwick tree too often and the net result is that code with fenwick tree is slower than code that uses for loop.