@caio 2018-11-04
Haha agreed! We'll have to keep this in mind when using the tree, but I can't think of a case where we would could potentially use it wrong (in this library scope, of course).

The code I see @ the honeycomb master branch is really old and doesn't have many of the fixes that happened over time (The fix #12, for example isn't there), so it ends up being faster because of their optimizations _and_ because it holds a lot less centroids. Maybe you could try the `ian.update-head` branch instead?