@caio 2018-11-04
Haha agreed about performance :) We'll have to keep it in mind, but doesn't seem risky.

Have you checked the `ian.update-head` branch of their fork? The master branch is pretty old and doesn't contain the fixes for #12 (which causes those negative values) and it holds a lot less centroids (which is one of the things that causes the bug heh). So it's a lot faster because of ian's optimizations and because it's inherently smaller.