@caio 2018-11-04
I see no disadvantages with this patch, should even help a bit with Merge-heavy workloads (there's a bit of discussion about one at #20) too. Merged, thanks a lot again!!