@caio 2018-11-04
For the record, I think (haven't verified) the fenwick tree change to use uint32 only works because our input is always sorted- if it could be unordered this would likely break:

```go
func (l *List) Append(n uint32) {
    i := len(l.tree)
    l.tree = append(l.tree, 0)
    l.tree[i] = n - l.Get(i)                                                                                                                                                                                
}
```