@vmihailenco 2018-11-04
I see. I guess that is fine as long as it improves performance :)

I've tried to use honeycomb fork and performance is 40% better for my workload (lots of merges) even with this patch. But either its binary serialization is incompatible or I am using it wrong but I get negative quantiles when I replace caio/go-tdigest with honeycomb/go-tdigest. So I guess I will continue to improve performance of this lib using ideas from honeycomb fork...